Moved platform specific header file to source.
